% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChooseLambda.R
\name{chooselambda}
\alias{chooselambda}
\title{How to choose tunning parameters in the process for selecting variables.}
\usage{
chooselambda(object, d)
}
\arguments{
\item{object}{object is the fitted lasso model.}

\item{d}{d is to measure the acceptance of standard deviation. For example, d=1 means we can only choose those lambda which are located in the range of minimum error with one standard deviation.d=2 means we can only choose those lambda which are located in the range of minimum error with two standard deviations.}
}
\value{
The lambda value.
}
\description{
How to choose tunning parameters in the process for selecting variables.
}
