% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearKernel.R
\name{Fun_Klinear}
\alias{Fun_Klinear}
\title{Calculate the Linear kernel for genomic matrix}
\usage{
Fun_Klinear(Gen)
}
\arguments{
\item{Gen}{Gen is a n*p genomic matrix with n subjects and p SNPs.}
}
\value{
The kernel matrix of the corresponding genomic matrix.
}
\description{
Calculate the Linear kernel for genomic matrix
}
