% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleKernel.R
\name{scaleK}
\alias{scaleK}
\title{Scaling the Linear kernel matrix for genomic matrix}
\usage{
scaleK(K)
}
\arguments{
\item{K}{K is the kernel matrix of the corresponding genomic matrix..}
}
\value{
The scaled kernel matrix.
}
\description{
Scaling the Linear kernel matrix for genomic matrix
}
